jobs-extract-data/json-transform/requirements.txt
boto3==1.19.9
botocore==1.22.9
numpy==1.21.2
pandas==1.3.3
PyYAML==6.0
py4j==0.10.9.2